Fix tree construction for vectors without `StaticArrays.setindex`
In #166 the `HyperRectangle` struct was changed to use the same vector
type as the input data for the mins and maxes, and to use
`StaticArrays.setindex` for "mutation". Since this function is not
implemented for all working input types (for example `Tensors.Vec`)
there was an `MethodError` thrown. This patch adds
`NearestNeighbors.setindex` which dispatches to `StaticArrays.setindex`
for `<: StaticArray.StaticArray` and for other types casts to `SVector`
and then back to the original type after mutating.